@@ -11,10 +11,10 @@ description: Frases 480 App
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.2.0+6
+version: 1.3.0+7
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -22,14 +22,14 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.2
-  http: ^0.12.0+2
-  flutter_spinkit: ^3.1.0
-  url_launcher: ^5.2.5
-  share: ^0.6.3+6
-  firebase_analytics: ^5.0.11
-  firebase_messaging: ^6.0.13
-  firebase_crashlytics: ^0.1.3+3
+  cupertino_icons: ^1.0.2
+  http: ^0.13.1
+  flutter_spinkit: ^5.0.0
+  url_launcher: ^6.0.3
+  share: ^2.0.1
+  firebase_analytics: ^8.0.1
+  firebase_messaging: ^9.1.2
+  firebase_crashlytics: ^2.0.1
 
 dev_dependencies:
   flutter_test:
